+++ apps/desktop/src/components/recast/index.ts
@@ -1,5 +1,6 @@
 export { default as ActionPanel } from "./ActionPanel.svelte";
 export { default as ConfirmDialog } from "./ConfirmDialog.svelte";
+export { default as PlayerDialog } from "./PlayerDialog.svelte";
 export { default as RecastCard } from "./RecastCard.svelte";
 export { default as RecastList } from "./RecastList.svelte";
 export { default as RecastRow } from "./RecastRow.svelte";